refactor(auth-commands): split login.rs into logical modules
Split login.rs into a login/ submodule tree: prompt (the credential-prompt
capability seam), host (the production Host provider), web_login (the
web-based login flow and its error), classic_login (the classic PUT flow,
add_user, and ClassicLoginOpError), and error (LoginError). The root keeps the
login orchestrator, LoginOptions, LoginHost, and the shared registry_join /
global_info / normalize_scope / safe_read_ini helpers, re-exporting the public
items. Pure code motion: no logic, signatures, error codes, or messages changed.
