+++ pacquet/crates/auth-commands/src/login.rs
@@ -26,172 +26,33 @@
 //! sit behind the `Sys` seam. See the "Dependency injection for tests" section
 //! of `pacquet/CODE_STYLE_GUIDE.md`.
 
-use std::{
-    future::Future,
-    io,
-    path::{Path, PathBuf},
-};
+use std::{io, path::Path};
 
-use derive_more::{Display, Error};
-use miette::Diagnostic;
-use pacquet_network::{ThrottledClient, encode_uri_component, nerf_dart, redact_and_sanitize};
+use pacquet_network::{ThrottledClient, nerf_dart, redact_and_sanitize};
 use pacquet_network_web_auth::{
-    Clock, EnterKeyListener, GenerateQrCodeError, Host as WebAuthHost, OpenUrl, OtpChallenge,
-    OtpError, PromptError, PromptOtp, Sleep, StdinIsTty, StdoutIsTty, SyntheticOtpError,
-    WebAuthFetch, WebAuthFetchError, WebAuthFetchOptions, WebAuthFetchResponse,
-    WebAuthRetryOptions, WebAuthTimeoutError, WebAuthTokenPollParams, WithOtpError,
-    generate_qr_code, poll_for_web_auth_token, prompt_browser_open, with_otp_handling,
+    Clock, EnterKeyListener, OpenUrl, PromptOtp, Sleep, StdinIsTty, StdoutIsTty, WebAuthFetch,
+    WebAuthFetchOptions, WebAuthRetryOptions,
 };
 use pacquet_reporter::{GlobalLog, LogEvent, LogLevel, Reporter};
-use serde_json::{Value, json};
 
 use crate::{
     ini::IniSettings,
     logout::{DEFAULT_REGISTRY, FsReadToString, FsWrite, normalize_registry_url},
 };
 
-/// Read a visible credential line — the `dialoguer::Input` builder chain (the
-/// username and email prompts). This is the blocking terminal read behind the
-/// capability seam, and nothing more: the production impl is that bare chain.
-///
-/// The surrounding algorithm — running the blocking read off the async runtime,
-/// selecting the visible or masked read by masking, and classifying a
-/// [`dialoguer::Error`] into a [`PromptError`] — deliberately does *not* live
-/// here or in [`PromptPassword`]. It lives in `prompt_line`, so each provider
-/// stays a bare builder chain while a test fakes only the read and that real
-/// algorithm still runs.
-pub trait PromptInput {
-    fn prompt_input(message: &str) -> Result<String, dialoguer::Error>;
-}
+mod classic_login;
+mod error;
+mod host;
+mod prompt;
+mod web_login;
 
-/// Read a masked credential secret — the `dialoguer::Password` builder chain
-/// (the password prompt). The masked counterpart of [`PromptInput`]; see there
-/// for how the surrounding algorithm is kept out of the provider.
-pub trait PromptPassword {
-    fn prompt_password(message: &str) -> Result<String, dialoguer::Error>;
-}
+pub use classic_login::ClassicLoginOpError;
+pub use error::LoginError;
+pub use host::Host;
+pub use prompt::{PromptInput, PromptPassword};
 
-/// Production provider for `pnpm login`. The credential prompts and `auth.ini`
-/// I/O are real; every OTP / web-authentication capability delegates to
-/// [`pacquet_network_web_auth::Host`], the shared production provider for that
-/// flow.
-pub struct Host;
-
-impl FsReadToString for Host {
-    fn read_to_string(path: &Path) -> io::Result<String> {
-        std::fs::read_to_string(path)
-    }
-}
-
-impl FsWrite for Host {
-    fn write(path: &Path, bytes: &[u8]) -> io::Result<()> {
-        pacquet_fs::write_atomic(path, bytes)
-    }
-}
-
-impl PromptInput for Host {
-    fn prompt_input(message: &str) -> Result<String, dialoguer::Error> {
-        dialoguer::Input::<String>::new().with_prompt(message).allow_empty(true).interact_text()
-    }
-}
-
-impl PromptPassword for Host {
-    fn prompt_password(message: &str) -> Result<String, dialoguer::Error> {
-        // `allow_empty_password(true)` mirrors `@inquirer/prompts` `password`,
-        // which returns an empty string on a bare Enter. Without it dialoguer
-        // loops until non-empty, so pnpm's empty-password path
-        // (`LOGIN_MISSING_CREDENTIALS`) would be unreachable.
-        dialoguer::Password::new().with_prompt(message).allow_empty_password(true).interact()
-    }
-}
-
-impl Clock for Host {
-    fn now_ms() -> u64 {
-        WebAuthHost::now_ms()
-    }
-}
-
-impl Sleep for Host {
-    fn sleep_ms(ms: u64) -> impl Future<Output = ()> {
-        WebAuthHost::sleep_ms(ms)
-    }
-}
-
-impl WebAuthFetch for Host {
-    fn fetch(
-        url: &str,
-        options: &WebAuthFetchOptions,
-    ) -> impl Future<Output = Result<WebAuthFetchResponse, WebAuthFetchError>> {
-        WebAuthHost::fetch(url, options)
-    }
-}
-
-impl StdinIsTty for Host {
-    fn stdin_is_tty() -> bool {
-        WebAuthHost::stdin_is_tty()
-    }
-}
-
-impl StdoutIsTty for Host {
-    fn stdout_is_tty() -> bool {
-        WebAuthHost::stdout_is_tty()
-    }
-}
-
-impl OpenUrl for Host {
-    fn open_url(url: &str) -> io::Result<()> {
-        WebAuthHost::open_url(url)
-    }
-}
-
-impl EnterKeyListener for Host {
-    type Handle = <WebAuthHost as EnterKeyListener>::Handle;
-    fn listen() -> io::Result<Self::Handle> {
-        WebAuthHost::listen()
-    }
-}
-
-impl PromptOtp for Host {
-    fn input(message: &str) -> impl Future<Output = Result<Option<String>, PromptError>> {
-        WebAuthHost::input(message)
-    }
-}
-
-/// Whether [`prompt_line`] reads a visible line or a masked secret.
-#[derive(Clone, Copy)]
-enum Masking {
-    Visible,
-    Masked,
-}
-
-/// Read one credential line off the async runtime (`dialoguer` is blocking):
-/// run the injected [`PromptInput`] read for [`Masking::Visible`] or the
-/// [`PromptPassword`] read for [`Masking::Masked`], then classify the outcome —
-/// an interrupted prompt (Ctrl-C) maps to [`PromptError::Cancelled`] (mirroring
-/// enquirer's `ExitPromptError`), a task panic or any other failure to
-/// [`PromptError::Other`].
-async fn prompt_line<Sys: PromptInput + PromptPassword + 'static>(
-    message: &str,
-    masking: Masking,
-) -> Result<String, PromptError> {
-    let message = message.to_owned();
-    tokio::task::spawn_blocking(move || match masking {
-        Masking::Visible => Sys::prompt_input(&message),
-        Masking::Masked => Sys::prompt_password(&message),
-    })
-    .await
-    .map_err(|join_error| PromptError::Other { reason: join_error.to_string() })?
-    .map_err(map_dialoguer_error)
-}
-
-fn map_dialoguer_error(error: dialoguer::Error) -> PromptError {
-    match error {
-        dialoguer::Error::IO(io) if io.kind() == io::ErrorKind::Interrupted => {
-            PromptError::Cancelled
-        }
-        dialoguer::Error::IO(io) => PromptError::Other { reason: io.to_string() },
-    }
-}
+use classic_login::classic_login;
+use web_login::{WebLoginFlowError, web_login};
 
 /// Inputs to [`login`]. The retry / timeout knobs come from pnpm's
 /// `fetchRetries` / `fetchTimeout` config and drive the web-auth poll.
@@ -335,224 +196,6 @@ fn safe_read_ini<Sys: FsReadToString>(path: &Path) -> Result<IniSettings, LoginE
     }
 }
 
-/// Drive the registry's web-based login: probe `-/v1/login`, then poll the
-/// returned `doneUrl` for the granted token while offering to open the login
-/// URL in a browser.
-async fn web_login<Sys, Reporter>(
-    http_client: &ThrottledClient,
-    registry: &str,
-    fetch_options: &WebAuthFetchOptions,
-) -> Result<String, WebLoginFlowError>
-where
-    Sys: Clock + Sleep + WebAuthFetch + StdinIsTty + EnterKeyListener + OpenUrl,
-    Reporter: self::Reporter,
-{
-    let login_url = registry_join(registry, "-/v1/login")
-        .map_err(|error| WebLoginFlowError::Transport { reason: error.to_string() })?;
-    let response = web_login_post(http_client, &login_url).await?;
-    if !response.ok {
-        return Err(WebLoginFlowError::Http { status: response.status, text: response.body });
-    }
-
-    let json = serde_json::from_str::<Value>(&response.body).unwrap_or(Value::Null);
-    let read = |field: &str| {
-        json.get(field).and_then(Value::as_str).filter(|value| !value.is_empty()).map(str::to_owned)
-    };
-    let (Some(auth_url), Some(done_url)) = (read("loginUrl"), read("doneUrl")) else {
-        return Err(WebLoginFlowError::InvalidResponse);
-    };
-
-    let qr_code = generate_qr_code(&auth_url).map_err(WebLoginFlowError::QrCode)?;
-    global_info::<Reporter>(&format!("Authenticate your account at:\n{auth_url}\n\n{qr_code}"));
-
-    let poll = poll_for_web_auth_token::<Sys>(WebAuthTokenPollParams {
-        done_url,
-        fetch_options: fetch_options.clone(),
-        timeout_ms: None,
-    });
-    prompt_browser_open::<Sys, Reporter, WebAuthTimeoutError, _>(&auth_url, poll)
-        .await
-        .map_err(WebLoginFlowError::Timeout)
-}
-
-/// Send the `POST -/v1/login` web-login probe and materialize its response.
-async fn web_login_post(
-    http_client: &ThrottledClient,
-    login_url: &str,
-) -> Result<HttpResponse, WebLoginFlowError> {
-    let guard = http_client.acquire_for_url(login_url).await;
-    let response = guard
-        .post(login_url)
-        .header("content-type", "application/json")
-        .header("accept", "application/json")
-        .header("npm-auth-type", "web")
-        .body("{}")
-        .send()
-        .await
-        .map_err(|error| WebLoginFlowError::Transport { reason: error.to_string() })?;
-    let ok = response.status().is_success();
-    let status = response.status().as_u16();
-    let body = response.text().await.unwrap_or_default();
-    Ok(HttpResponse { ok, status, body })
-}
-
-/// Prompt for username / password / email and register the user through
-/// `PUT -/user/org.couchdb.user:<name>`, satisfying an OTP challenge if the
-/// registry raises one.
-async fn classic_login<Sys, Reporter>(
-    http_client: &ThrottledClient,
-    registry: &str,
-    fetch_options: WebAuthFetchOptions,
-) -> Result<String, LoginError>
-where
-    Sys: Clock
-        + Sleep
-        + WebAuthFetch
-        + StdinIsTty
-        + StdoutIsTty
-        + EnterKeyListener
-        + OpenUrl
-        + PromptOtp
-        + PromptInput
-        + PromptPassword
-        + 'static,
-    Reporter: self::Reporter,
-{
-    let username = read_credential(prompt_line::<Sys>("Username:", Masking::Visible)).await?;
-    let password = read_credential(prompt_line::<Sys>("Password:", Masking::Masked)).await?;
-    let email =
-        read_credential(prompt_line::<Sys>("Email (this IS public):", Masking::Visible)).await?;
-
-    if username.is_empty() || password.is_empty() || email.is_empty() {
-        return Err(LoginError::MissingCredentials);
-    }
-
-    let username_ref = username.as_str();
-    let password_ref = password.as_str();
-    let email_ref = email.as_str();
-    let token = with_otp_handling::<Sys, Reporter, String, ClassicLoginOpError, _, _>(
-        fetch_options,
-        // A plain `FnMut` returning an `async move` block: the future is a
-        // concrete type carrying only `Copy` borrows and the by-value OTP, so
-        // it borrows nothing from the closure — see `with_otp_handling`'s
-        // `Operation` bound.
-        move |otp: Option<String>| async move {
-            add_user(http_client, registry, username_ref, password_ref, email_ref, otp.as_deref())
-                .await
-                .map_err(add_user_error_to_op::<Reporter>)
-        },
-    )
-    .await
-    .map_err(LoginError::ClassicLogin)?;
-
-    global_info::<Reporter>(&format!("Logged in as {username}"));
-
-    Ok(token)
-}
-
-/// Await a credential prompt, mapping an aborted prompt to
-/// [`LoginError::Canceled`] and any other failure to [`LoginError::Prompt`].
-async fn read_credential(
-    prompt: impl Future<Output = Result<String, PromptError>>,
-) -> Result<String, LoginError> {
-    match prompt.await {
-        Ok(value) => Ok(value),
-        Err(PromptError::Cancelled) => Err(LoginError::Canceled),
-        Err(error) => Err(LoginError::Prompt { reason: error.to_string() }),
-    }
-}
-
-/// Register a user via `PUT -/user/org.couchdb.user:<name>`, returning the
-/// granted token. `otp` populates the `npm-otp` header on the retry pass.
-async fn add_user(
-    http_client: &ThrottledClient,
-    registry: &str,
-    username: &str,
-    password: &str,
-    email: &str,
-    otp: Option<&str>,
-) -> Result<String, AddUserError> {
-    let url = registry_join(
-        registry,
-        &format!("-/user/org.couchdb.user:{}", encode_uri_component(username)),
-    )
-    .map_err(|error| AddUserError::Transport { reason: error.to_string() })?;
-    let document = json!({
-        "_id": format!("org.couchdb.user:{username}"),
-        "name": username,
-        "password": password,
-        "email": email,
-        "type": "user",
-    });
-    let body = serde_json::to_string(&document).expect("serialize addUser document");
-
-    let guard = http_client.acquire_for_url(&url).await;
-    let mut request = guard
-        .put(&url)
-        .header("content-type", "application/json")
-        .header("accept", "application/json")
-        .header("npm-auth-type", "web")
-        .body(body);
-    if let Some(otp) = otp {
-        request = request.header("npm-otp", otp);
-    }
-
-    let response = request
-        .send()
-        .await
-        .map_err(|error| AddUserError::Transport { reason: error.to_string() })?;
-    let ok = response.status().is_success();
-    let status = response.status().as_u16();
-    // Join every `WWW-Authenticate` header the way the Fetch `Headers.get`
-    // pnpm relies on does, so an `otp` challenge that isn't the first of
-    // several challenge headers is still detected.
-    let www_authenticate = {
-        let joined = response
-            .headers()
-            .get_all(reqwest::header::WWW_AUTHENTICATE)
-            .iter()
-            .filter_map(|value| value.to_str().ok())
-            .collect::<Vec<_>>()
-            .join(", ");
-        (!joined.is_empty()).then_some(joined)
-    };
-    let text = response.text().await.unwrap_or_default();
-
-    if !ok {
-        return Err(AddUserError::Http { status, text, www_authenticate });
-    }
-
-    match serde_json::from_str::<Value>(&text)
-        .ok()
-        .and_then(|json| json.get("token").and_then(Value::as_str).map(str::to_owned))
-    {
-        Some(token) if !token.is_empty() => Ok(token),
-        _ => Err(AddUserError::NoToken),
-    }
-}
-
-/// Classify an [`add_user`] failure for [`with_otp_handling`]: a `401` that
-/// advertises `otp` in `WWW-Authenticate` becomes an OTP challenge (its body
-/// parsed for `authUrl` / `doneUrl`), everything else a terminal error.
-fn add_user_error_to_op<Reporter: self::Reporter>(error: AddUserError) -> ClassicLoginOpError {
-    match error {
-        // Mirrors pnpm's `err.responseHeaders.get('www-authenticate')?.includes('otp')`.
-        AddUserError::Http { status, text, www_authenticate }
-            if status == 401
-                && www_authenticate.as_deref().is_some_and(|value| value.contains("otp")) =>
-        {
-            let json = serde_json::from_str::<Value>(&text).ok();
-            let challenge = SyntheticOtpError::from_unknown_body::<Reporter>(json.as_ref())
-                .as_otp_challenge()
-                .expect("SyntheticOtpError is always an OTP challenge");
-            ClassicLoginOpError::Otp { challenge }
-        }
-        AddUserError::Http { status, text, .. } => ClassicLoginOpError::Failed { status, text },
-        AddUserError::NoToken => ClassicLoginOpError::NoToken,
-        AddUserError::Transport { reason } => ClassicLoginOpError::Transport { reason },
-    }
-}
-
 /// Resolve `path` against the registry the way `new URL(path, registry)` does.
 fn registry_join(registry: &str, path: &str) -> Result<String, url::ParseError> {
     url::Url::parse(registry)?.join(path).map(String::from)
@@ -563,159 +206,6 @@ fn global_info<Reporter: self::Reporter>(message: &str) {
         level: LogLevel::Info,
         message: message.to_owned(),
     }));
-}
-
-/// A materialized registry response — the fields [`web_login`] and
-/// [`add_user`] read after the request completes.
-struct HttpResponse {
-    ok: bool,
-    status: u16,
-    body: String,
-}
-
-/// Failure surface of the web-based login flow. Internal to this module: the
-/// caller either falls back to classic login (on [`Http`](Self::Http) 404/405)
-/// or maps the rest into a [`LoginError`] via [`From`].
-enum WebLoginFlowError {
-    /// The web-login probe returned a non-success status.
-    Http { status: u16, text: String },
-    /// The probe succeeded but the body lacked a usable `loginUrl` / `doneUrl`.
-    InvalidResponse,
-    /// The token poll exceeded its budget.
-    Timeout(WebAuthTimeoutError),
-    /// The login URL could not be rendered as a QR code.
-    QrCode(GenerateQrCodeError),
-    /// The probe request never produced a response.
-    Transport { reason: String },
-}
-
-impl From<WebLoginFlowError> for LoginError {
-    fn from(error: WebLoginFlowError) -> Self {
-        match error {
-            WebLoginFlowError::Http { status, text } => LoginError::WebLoginFailed { status, text },
-            WebLoginFlowError::InvalidResponse => LoginError::InvalidResponse,
-            WebLoginFlowError::Timeout(timeout) => LoginError::WebAuthTimeout(timeout),
-            WebLoginFlowError::QrCode(qr) => LoginError::QrCode(qr),
-            WebLoginFlowError::Transport { reason } => LoginError::Request { reason },
-        }
-    }
-}
-
-/// Failure surface of [`add_user`]. Internal to this module: mapped into a
-/// [`ClassicLoginOpError`] for [`with_otp_handling`].
-enum AddUserError {
-    /// The registry returned a non-success status.
-    Http { status: u16, text: String, www_authenticate: Option<String> },
-    /// The registry accepted the request but returned no token.
-    NoToken,
-    /// The request never produced a response.
-    Transport { reason: String },
-}
-
-/// The error the classic-login operation hands to [`with_otp_handling`]. The
-/// [`Otp`](Self::Otp) arm is the only challenge it acts on; the rest are
-/// terminal and surface (transparently) as the login result's error, carrying
-/// pnpm's stable codes.
-#[derive(Debug, Display, Error, Diagnostic)]
-pub enum ClassicLoginOpError {
-    #[display("The registry requires a one-time password to complete the login")]
-    #[diagnostic(code(pacquet_auth_commands::login_otp_required))]
-    Otp {
-        #[error(not(source))]
-        challenge: OtpChallenge,
-    },
-
-    #[display("Login failed (HTTP {status}): {text}")]
-    #[diagnostic(code(ERR_PNPM_LOGIN_FAILED))]
-    Failed { status: u16, text: String },
-
-    #[display("The registry did not return an authentication token")]
-    #[diagnostic(code(ERR_PNPM_LOGIN_NO_TOKEN))]
-    NoToken,
-
-    #[display("The login request failed: {reason}")]
-    #[diagnostic(code(pacquet_auth_commands::login_request_failed))]
-    Transport {
-        #[error(not(source))]
-        reason: String,
-    },
-}
-
-impl OtpError for ClassicLoginOpError {
-    fn as_otp_challenge(&self) -> Option<OtpChallenge> {
-        match self {
-            ClassicLoginOpError::Otp { challenge } => Some(challenge.clone()),
-            _ => None,
-        }
-    }
-}
-
-/// Errors surfaced by [`login`]. The user-facing variants carry pnpm's stable
-/// error codes and messages verbatim; the classic-login and web-auth arms
-/// surface their inner diagnostics transparently.
-#[derive(Debug, Display, Error, Diagnostic)]
-pub enum LoginError {
-    #[display("The login command requires an interactive terminal")]
-    #[diagnostic(code(ERR_PNPM_LOGIN_NON_INTERACTIVE))]
-    NonInteractive,
-
-    #[display("The registry returned an invalid response for web-based login")]
-    #[diagnostic(code(ERR_PNPM_LOGIN_INVALID_RESPONSE))]
-    InvalidResponse,
-
-    #[display("Username, password, and email are all required")]
-    #[diagnostic(code(ERR_PNPM_LOGIN_MISSING_CREDENTIALS))]
-    MissingCredentials,
-
-    #[display("Login canceled")]
-    #[diagnostic(code(ERR_PNPM_LOGIN_CANCELED))]
-    Canceled,
-
-    #[display("Web-based login failed (HTTP {status}): {text}")]
-    #[diagnostic(code(ERR_PNPM_WEB_LOGIN_FAILED))]
-    WebLoginFailed { status: u16, text: String },
-
-    #[display("{_0}")]
-    #[diagnostic(transparent)]
-    ClassicLogin(WithOtpError<ClassicLoginOpError>),
-
-    #[display("{_0}")]
-    #[diagnostic(transparent)]
-    WebAuthTimeout(WebAuthTimeoutError),
-
-    #[display("Failed to render the login QR code: {_0}")]
-    #[diagnostic(code(pacquet_auth_commands::login_qr_code))]
-    QrCode(#[error(source)] GenerateQrCodeError),
-
-    #[display("The login request failed: {reason}")]
-    #[diagnostic(code(pacquet_auth_commands::login_request_failed))]
-    Request {
-        #[error(not(source))]
-        reason: String,
-    },
-
-    #[display("Failed to read the login prompt: {reason}")]
-    #[diagnostic(code(pacquet_auth_commands::login_prompt_failed))]
-    Prompt {
-        #[error(not(source))]
-        reason: String,
-    },
-
-    #[display("Failed to read auth.ini at {}: {error}", path.display())]
-    #[diagnostic(code(pacquet_auth_commands::read_auth_ini))]
-    ReadAuthIni {
-        path: PathBuf,
-        #[error(source)]
-        error: io::Error,
-    },
-
-    #[display("Failed to write auth.ini at {}: {error}", path.display())]
-    #[diagnostic(code(pacquet_auth_commands::write_auth_ini))]
-    WriteAuthIni {
-        path: PathBuf,
-        #[error(source)]
-        error: io::Error,
-    },
 }
 
 #[cfg(test)]
