+++ pacquet/crates/auth-commands/src/login/test_support.rs
@@ -16,7 +16,7 @@ use crate::ini::IniSettings;
 
 /// A scripted response for a credential prompt, keyed on the prompt message.
 /// The error half is `dialoguer::Error` — exactly what the real terminal read
-/// yields — so a script can drive [`super::prompt_line`]'s real error
+/// yields — so a script can drive [`super::prompt::prompt_line`]'s real error
 /// classification (e.g. an interrupt mapping to a canceled login). `Send`
 /// because `prompt_line` calls the fake read from a `spawn_blocking` thread.
 pub(crate) type PromptScript = Box<dyn FnMut(&str) -> Result<String, dialoguer::Error> + Send>;
